+++ src/l6470/motors.cpp
@@ -136,7 +136,7 @@ void Motors::setUp(){
 	//this->setFullSpeed(120);       // microstep below 120 steps/s
 }
 
-void Motors::setMaxSpeedForBoth(uint8_t speed){
+void Motors::setMaxSpeedForBoth(float speed){
 	m_nPosition = 0;
 	this->setMaxSpeed(speed);
 	m_nPosition = 1;
